requirements.txt
Django==1.5.4
Sphinx>=1.1.2
Werkzeug==0.8.3
django-debug-toolbar==0.9.4

mock==1.0.1
pytest==2.4.2
pytest-django==2.3.1
pytest-sugar==0.2.2
pytest-cov==1.6
django-oscar-testsupport==0.2
django-dynamic-fixture==1.4.2
WebTest==1.3.4
django-webtest==1.5.3

factory-boy==2.2.1
humanize==0.5

docutils==0.8.1
flake8==0.8
pyflakes==0.5.0
coverage==3.5.1

-e git+https://github.com/tangentlabs/django-oscar.git#egg=django-oscar

# required for tastypie
mimeparse==0.1.3
django-haystack==2.1.0
